added the verification of a merkle path
